Changing codeboxes in the Copy Constraints section
Codebox 11:

	1. There were a few things mixed up in the code boxes. I corrected it while trying to preserve the structure. If this PR goes through, I would also like to change the variable names to be more precise because if we want to have the entire protocol things will get very confusing

	2. find_permutation() currently just returns a polynomial not a permutation. I did send another PR to change that. I generate the permutation in the next codebox.

	3. We actually want to create the polynomials and not the permutation so I guess we should use gen_poly()

	4. We want 3  polynomials combined described in the text above the box as witness_x_1.  Those return the identity domain: a: [0,..5], b: [6,..,11], c: [12,..,16] respectively when evaluated over [0,5]

	5. The polynomial described in the text as witness_x_2 from the text is produced in the next codebox below in what you describe as copy_constraints.

	6. I think it is actually very important to also note here that we dont need the witness to generate these polynomials. It is enough to know what wires are duplicated which you can see from the circuit. These polynomials are in practice generated in the setup phase.

Codebox 12:
In addition to witness_y we need a polynomial that evaluates to the permuted witness over [0,…,15] because we want to check that the witness evaluated over [0,…,15] is the same as the permuted witness when evaluated over the permuted indices.

Changes to Poly.py:
We need to be able to change the domain so I enabled using x's
Note that we are currently not using field elements as coefficients. Its fine for this example but we wont be able to go very far.

Changes to copy constraints:
I guess there are better algorithms to shuffle the indices but that should work.

I didnt have a chance yet to go through the snake-charmers-tactical-guide. Sorry in advance for any unconvential things I did.